% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialQC.R
\name{addBasemapLayers}
\alias{addBasemapLayers}
\title{addBasemapLayers
This file adds a number of spatial objects to the output gpkg file, such as
landing locations, offshore banks, surfclam fishing areas, NAFO areas, and 
bathymetry.}
\usage{
addBasemapLayers(...)
}
\arguments{
\item{...}{Additional arguments passed on to other functions.}
}
\value{
This function does not return a value. It writes layers to a gpkg file which is output to a specified directory.
}
\description{
addBasemapLayers
This file adds a number of spatial objects to the output gpkg file, such as
landing locations, offshore banks, surfclam fishing areas, NAFO areas, and 
bathymetry.
}
