% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{any_date_time}
\alias{any_date_time}
\title{any_date_time
This function checks if x is a date or a time object.}
\usage{
any_date_time(x)
}
\arguments{
\item{x}{This is an object to check for whether or not it is a date or a time object}
}
\value{
A logical value indicating whether the input object is a date or datetime object.
}
\description{
any_date_time
This function checks if x is a date or a time object.
}
