% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{banks_sf}
\alias{banks_sf}
\alias{SurfClamFAs_sf}
\title{SurfClamFAs_sf Metadata}
\format{
sf polygon
\describe{
\item{Name}{Name of the area}
}
}
\usage{
SurfClamFAs_sf
}
\description{
An sf polygon for the existing Surf clam fishing areas.  The Grand Bank area is just NAFO areas 
3LNO.

#' * This object was copied from Maritimes/Mar.data::coast_lores_sf on Aug 14, 2024 to prevent having
to add a dependency
}
\details{
The CRS of this data is EPSG:4326 (WGS 84).
}
\keyword{datasets}
