% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{check_field_values}
\alias{check_field_values}
\title{check_field_values
This is a generic checking function that verifies that values within \code{chkDf$field}
are within \code{acceptable_values} (which can either be a complete list, or if \code{useRange=TRUE}, the bounding values of a range).}
\usage{
check_field_values(
  chkDf = NULL,
  field = NULL,
  acceptable_values = NULL,
  output_messages = NULL,
  show.passed.checks = NULL,
  issues = NULL,
  lineSep = "====================================",
  IDListOnly = FALSE,
  na_acceptable = FALSE,
  useRange = FALSE,
  suppl_out = "",
  tab_suppl = FALSE,
  ...
)
}
\arguments{
\item{chkDf}{the default is \code{NULL}.  This is the dataframe object being submitted to this function.}

\item{field}{default is \code{FALSE} This is the name of the field to check}

\item{acceptable_values}{default is \code{NULL} This is a vector of values that are acceptable (or, if \code{useRange=T}, the min and max values for an acceptable range)}

\item{output_messages}{default is \code{NULL} This is a text block of messages.  If one is supplied, new messages will be appended to it.}

\item{show.passed.checks}{default is \code{NULL} If output is overwhelming, this can be set to FALSE, and only failing checks will be displayed.}

\item{issues}{default is \code{NULL} This is a counter of discovered issues.  It will be incremented as issues are found.}

\item{lineSep}{default is \code{"===================================="}.  This is text that will be added to the output between different checks.}

\item{IDListOnly}{default is \code{FALSE} Tabular data can be returned, but setting this to TRUE ensures that only the IDs of problematic records is returned.}

\item{na_acceptable}{default is \code{FALSE} This specifies whether or not values of NA are allowed.  If \code{TRUE}, the user will not be alerted when NAs are found.}

\item{useRange}{default is \code{FALSE} This specifies whether or not \code{acceptable_values} specifies discrete values or the bounds of a range}

\item{suppl_out}{default is \code{""} This allows submission of additional details about the test that will be included in the output (e.g. " (i.e. all weights are between 3000 and 6000 kg)")}

\item{tab_suppl}{default is \code{FALSE} This can be set to TRUE in order to add an additional tab in front of the output.  This can facilitate readability of the output.}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
list of output messages and an incremented value for \code{issues}.
}
\description{
check_field_values
This is a generic checking function that verifies that values within \code{chkDf$field}
are within \code{acceptable_values} (which can either be a complete list, or if \code{useRange=TRUE}, the bounding values of a range).
}
