% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCaptain.R
\name{checkCaptain}
\alias{checkCaptain}
\title{checkCaptain
 This function helps verify that the reported captains are valid for  the vessels 
 they are piloting}
\usage{
checkCaptain(chkDf = NULL, output_messages = NULL, issues = NULL, ...)
}
\arguments{
\item{chkDf}{default is \code{NULL}.  This is the dataframe object being submitted to this function.}

\item{output_messages}{default is \code{NULL}}

\item{issues}{default is \code{NULL}}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
list
}
\description{
checkCaptain
 This function helps verify that the reported captains are valid for  the vessels 
 they are piloting
}
