% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamCompLogCW.R
\name{clamCompLogCW}
\alias{clamCompLogCW}
\title{clamCompLogCW
This function compares the contents of the loaded CW and Log files to verify that
the positions correspond, and that the reported weights of the carious products
match between the 2 files.}
\usage{
clamCompLogCW(clamCW = NULL, clamLog = NULL, ...)
}
\arguments{
\item{clamCW}{default is \code{NULL}}

\item{clamLog}{default is \code{NULL}}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
list
}
\description{
clamCompLogCW
This function compares the contents of the loaded CW and Log files to verify that
the positions correspond, and that the reported weights of the carious products
match between the 2 files.
}
