% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamCW_QC.R
\name{clamCW_QC}
\alias{clamCW_QC}
\title{clamCW_QC}
\usage{
clamCW_QC(cwFolder = NULL, ...)
}
\arguments{
\item{cwFolder}{default is \code{NULL}}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
This function does not return a value. It writes output files to the specified directory.
}
\description{
clamCW_QC
}
