% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamLog_QC.R
\name{clamLog_QC}
\alias{clamLog_QC}
\title{clamLog_QC}
\usage{
clamLog_QC(logFolder = NULL, ...)
}
\arguments{
\item{logFolder}{default is \code{NULL}}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
data.frame
}
\description{
clamLog_QC
}
