% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamLog_QC_file.R
\name{clamLog_QC_file}
\alias{clamLog_QC_file}
\title{clamLog_QC_file
This function extracts and reformats the contents of xlsx Log files so they 
can be compared by other functions within this package.}
\usage{
clamLog_QC_file(file = NULL, output = NULL, issues = NULL, ...)
}
\arguments{
\item{file}{default is \code{NULL}.  This is the path to a Log (xlsx) file.}

\item{output}{default is \code{NULL}}

\item{issues}{default is \code{NULL}}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
data.frame
}
\description{
clamLog_QC_file
This function extracts and reformats the contents of xlsx Log files so they 
can be compared by other functions within this package.
}
\examples{
\dontrun{
logFile <- clamLog_QC_file(file ="C:/R Cdn Clam DFO Log- Jan-5-8 2023.xlsx" )
} 
}
