% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clamQC.R
\name{clamQC}
\alias{clamQC}
\title{clamQC}
\usage{
clamQC(cwFolder = NULL, logFolder = NULL, resultsFolder = NULL, ...)
}
\arguments{
\item{cwFolder}{default is \code{NULL} This is a folder containing all of the various files submitted (e.g. Commercial_Sample_Profile.xlsx, Discard.xlsx, Gear.xlsx, Length_Frequency.xlsx, Products.xlsx, etc)}

\item{logFolder}{default is \code{NULL} This is a folder of the weekly onboard log data (e.g. "AE Cdn Clam DFO Log Aug 3-Aug 6.xlsx")}

\item{resultsFolder}{default is \code{NULL} This is an output folder, where the package will drop results}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
This function does not return a value. It writes output files to the specified directory.
}
\description{
clamQC
}
\examples{
\dontrun{
test <- clamQC(cwFolder = "C:/Original CW Log Data/2023/2023BC Q1&Q2/",
logFolder = "C:/HailedOnboardLogData/2023/BC/",
resultsFolder = "C:/exampleOutput/", 
cxn = <oracle connection>>, 
offline=F, debug=F,
gpkgName = "example.gpkg")
}
}
