% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{common_path}
\alias{common_path}
\title{common_path
Given 2 folders, this function will find the common path between the two.  
This was created for use with windows.}
\usage{
common_path(path1, path2)
}
\arguments{
\item{path1}{This is a path}

\item{path2}{This is a different path}
}
\value{
a string
}
\description{
common_path
Given 2 folders, this function will find the common path between the two.  
This was created for use with windows.
}
