% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CW_check_merge.R
\name{CW_check_merge}
\alias{CW_check_merge}
\title{CW_check_merge}
\usage{
CW_check_merge(
  df1 = NULL,
  df2 = NULL,
  df1_name = NULL,
  df2_name = NULL,
  output_gen = NULL,
  issues_gen = NULL,
  ...
)
}
\arguments{
\item{df1}{default is \code{NULL}.  This is the first dataframe object being submitted to this function.}

\item{df2}{default is \code{NULL}.  This is the second dataframe object being submitted to this function.}

\item{df1_name}{default is \code{NULL}.  This is the name of \code{df1}.}

\item{df2_name}{default is \code{NULL}.  This is the name of \code{df2}.}

\item{output_gen}{default is \code{NULL}.}

\item{issues_gen}{default is \code{NULL}.}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
list containing output messages and a count of the number of issues found
}
\description{
CW_check_merge
}
