% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CW_fields_QC.R
\name{CW_fields_QC}
\alias{CW_fields_QC}
\title{CW_fields_QC}
\usage{
CW_fields_QC(file = NULL, ...)
}
\arguments{
\item{file}{default is \code{NULL}.  This is the path to a CW-submitted file.}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
list
}
\description{
CW_fields_QC
}
