% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CW_generic_QC.R
\name{CW_generic_QC}
\alias{CW_generic_QC}
\title{CW_generic_QC
This function runs a number of checks to ensure that values from the CW files
are within acceptable ranges}
\usage{
CW_generic_QC(df = NULL, output_messages = NULL, issues = NULL, ...)
}
\arguments{
\item{df}{default is \code{NULL}.  This is the dataframe object being submitted to this function.}

\item{output_messages}{default is \code{NULL}}

\item{issues}{default is \code{NULL}}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
list
}
\description{
CW_generic_QC
This function runs a number of checks to ensure that values from the CW files
are within acceptable ranges
}
