% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CW_handleRecord.R
\name{CW_handleRecord}
\alias{CW_handleRecord}
\title{CW_handleRecord}
\usage{
CW_handleRecord(chkDf = NULL, layer_name = NULL, ...)
}
\arguments{
\item{chkDf}{default is \code{NULL}.  This is the dataframe object being submitted to this function.}

\item{layer_name}{default is \code{NULL}}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
list
}
\description{
CW_handleRecord
}
