% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{detectFishingCW}
\alias{detectFishingCW}
\title{detectFishingCW
This function uses various CW fields to verify whether or not fishing was occurring.}
\usage{
detectFishingCW(df = NULL)
}
\arguments{
\item{df}{default is \code{NULL}.  This is the dataframe object being submitted to this function.}
}
\value{
dataframe
}
\description{
detectFishingCW
This function uses various CW fields to verify whether or not fishing was occurring.
}
