% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{detectFishingLOG}
\alias{detectFishingLOG}
\title{detectFishingLOG
This function uses various Log fields to verify whether or not fishing was occurring.}
\usage{
detectFishingLOG(df = NULL)
}
\arguments{
\item{df}{default is \code{NULL}.  This is the dataframe object being submitted to this function.}
}
\value{
dataframe
}
\description{
detectFishingLOG
This function uses various Log fields to verify whether or not fishing was occurring.
}
