% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{df.diff}
\alias{df.diff}
\title{df.diff
This function find records that are different between 2 identically formatted 
data frames.}
\usage{
df.diff(df1, df2)
}
\arguments{
\item{df1}{This is the first dataframe object being submitted to this function.}

\item{df2}{This is the second dataframe object being submitted to this function.}
}
\value{
dataframe
}
\description{
df.diff
This function find records that are different between 2 identically formatted 
data frames.
}
