% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{ETOPO_1_Contours100_sf}
\alias{ETOPO_1_Contours100_sf}
\title{ETOPO_1_Contours100_sf Metadata}
\format{
sf linestring
\describe{
\item{ID}{ESRI ArcMap Field}
\item{CONTOUR}{ESRI ArcMap Field}
}
}
\usage{
ETOPO_1_Contours100_sf
}
\description{
This file was imported as an sf object from "R:/Science/Population Ecology Division/Sections/
Eastern Scotian Shelf/Clam Group/Maps/FishingAreas/shp/ETOPO_1_Contours100.shp" on Aug 14, 2024
}
\details{
The CRS of this data is EPSG:4326 (WGS 84).
}
\keyword{datasets}
