% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{get_global_params}
\alias{get_global_params}
\title{get_global_params
This function is used to set global parameters that can be used by other functions.}
\usage{
get_global_params()
}
\value{
list
}
\description{
get_global_params
This function is used to set global parameters that can be used by other functions.
}
