% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{getCreds}
\alias{getCreds}
\title{getCreds
This function just provides a reusable mechanism for working with Oracle credentials}
\usage{
getCreds(...)
}
\arguments{
\item{...}{Additional arguments passed on to other functions.}
}
\value{
list
}
\description{
getCreds
This function just provides a reusable mechanism for working with Oracle credentials
}
