% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{landingLocations_sf}
\alias{landingLocations_sf}
\title{landingLocations_sf Metadata}
\format{
sf point
\describe{
\item{Location}{Name of the location}
}
}
\usage{
landingLocations_sf
}
\description{
An sf point file showing the positions of Mulgrave and Argentia
}
\details{
The CRS of this data is EPSG:4326 (WGS 84).
}
\keyword{datasets}
