% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{override_params}
\alias{override_params}
\title{override_params
In the event that a submitted parameter has the same name as a global parameter, 
this function ensures that the submitted parameter is retained.}
\usage{
override_params(func_params)
}
\arguments{
\item{func_params}{This is a list of parameters that were submitted to a function  - typically via \code{func_params <- list(...)}}
}
\value{
list
}
\description{
override_params
In the event that a submitted parameter has the same name as a global parameter, 
this function ensures that the submitted parameter is retained.
}
