% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_utils.R
\name{sanitize_input}
\alias{sanitize_input}
\title{sanitize_input
This function removes trailing slashes from submitted strings}
\usage{
sanitize_input(input)
}
\arguments{
\item{input}{This is a string (typically a path)}
}
\value{
A string with any trailing slashes removed
}
\description{
sanitize_input
This function removes trailing slashes from submitted strings
}
