% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialQC.R
\name{spatialQC}
\alias{spatialQC}
\title{spatialQC
This function takes a dataframe with coordinates, and for each position, checks
where they are relative to:
1 - NAFO subunits (Mar.data::NAFOSubunits_sf)
2 - Surfclam Fishing areas (Mar.data::SurfClamFAs_sf)
3 - Major offshore banks on the Scotian Shelf (Mar.data::banks_sf)
4 - Depth contour min and max (Mar.data::Bathy_sf)}
\usage{
spatialQC(chkDf, layerName = "layer", ...)
}
\arguments{
\item{chkDf}{default is \code{NULL}.  This is the dataframe object being submitted to this function.}

\item{layerName}{default is \code{"layer"}}

\item{...}{Additional arguments passed on to other functions.}
}
\value{
This function returns an sf object corresponding with the submitted 
df, but with several extra fields.  It also generates a new layer within a gpkg file.
}
\description{
spatialQC
This function takes a dataframe with coordinates, and for each position, checks
where they are relative to:
1 - NAFO subunits (Mar.data::NAFOSubunits_sf)
2 - Surfclam Fishing areas (Mar.data::SurfClamFAs_sf)
3 - Major offshore banks on the Scotian Shelf (Mar.data::banks_sf)
4 - Depth contour min and max (Mar.data::Bathy_sf)
}
